+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [] -
+- fix downloading dataset for tournaments > 1
+
 ## [1.6.0] - 2019-07-10
 - default to tournament 8 `katzuagi`
 - update docstring